requirements.txt
pymupdf
pillow
imagehash
pytesseract
